requirements.txt
streamlit
python-dotenv
google-genai
google-generativeai
